workout refactor - workoutUnit

+++ LiftLogixServer/src/main/java/com/liftlogix/services/PersonalPlanService.java
@@ -4,9 +4,9 @@ import com.liftlogix.convert.BasicPersonalPlanDTOMapper;
 import com.liftlogix.convert.PersonalPlanDTOMapper;
 import com.liftlogix.dto.*;
 import com.liftlogix.exceptions.AuthorizationException;
-import com.liftlogix.models.Coach;
-import com.liftlogix.models.PersonalPlan;
-import com.liftlogix.models.User;
+import com.liftlogix.exceptions.NoActivePlanException;
+import com.liftlogix.models.*;
+import com.liftlogix.repositories.ClientRepository;
 import com.liftlogix.repositories.PersonalPlanRepository;
 import com.liftlogix.types.Role;
 import jakarta.persistence.EntityNotFoundException;
@@ -23,27 +23,44 @@ import java.util.stream.Collectors;
 @AllArgsConstructor
 public class PersonalPlanService {
     private final PersonalPlanRepository personalPlanRepository;
+    private final ClientRepository clientRepository;
     private final PersonalPlanDTOMapper personalPlanDTOMapper;
     private final BasicPersonalPlanDTOMapper basicPersonalPlanDTOMapper;
+    private final CoachSchedulerService coachSchedulerService;
 
-    public List<BasicPersonalPlanDTO> getAllClientPlans(Long clientId) {
+    public List<BasicPersonalPlanDTO> getAllClientPlans(Long clientId, User user) {
         List<PersonalPlan> personalPlans = personalPlanRepository.findByClientId(clientId);
+
+        if (!personalPlans.isEmpty()) {
+            PersonalPlan plan = personalPlans.getFirst();
+            Client client = plan.getClient();
+
+            if (!Objects.equals(client.getCoach().getEmail(), user.getEmail()) && !user.getRole().equals(Role.ADMIN)) {
+                throw new AuthorizationException("You are not authorized");
+            }
+        }
 
         return personalPlans.stream()
                 .map(basicPersonalPlanDTOMapper::mapEntityToDTO)
                 .collect(Collectors.toList());
     }
 
-    public PersonalPlanDTO getPlanDetails(Long id) {
+    public PersonalPlanDTO getPlanDetails(Long id, User user) {
         PersonalPlan plan = personalPlanRepository.findById(id).orElseThrow(
                 () -> new EntityNotFoundException("Personal plan not found")
         );
+
+        Client client = plan.getClient();
+
+        if (!Objects.equals(client.getCoach().getEmail(), user.getEmail()) && !user.getRole().equals(Role.ADMIN)) {
+            throw new AuthorizationException("You are not authorized");
+        }
 
         PersonalPlanDTO planDTO = personalPlanDTOMapper.mapEntityToDTO(plan);
 
         for (MesocycleDTO mesocycle : planDTO.getMesocycles()) {
             for (MicrocycleDTO microcycle : mesocycle.getMicrocycles()) {
-                for (WorkoutDTO workout : microcycle.getWorkouts()) {
+                for (WorkoutUnitDTO workout : microcycle.getWorkoutUnits()) {
                     workout.getWorkoutExercises().sort(Comparator.comparingLong(WorkoutExerciseDTO::getId));
                 }
             }
@@ -52,32 +69,64 @@ public class PersonalPlanService {
         return planDTO;
     }
 
-    public Optional<PersonalPlanDTO> getActivePlanByClientId(Long clientId) {
+    public PersonalPlanDTO getActivePlanByClientId(Long clientId, User user) {
+        Client client = clientRepository.findById(clientId).orElseThrow(
+                () -> new EntityNotFoundException("Client not found")
+        );
+
+        if (!Objects.equals(client.getCoach().getEmail(), user.getEmail()) && !user.getRole().equals(Role.ADMIN)) {
+            throw new AuthorizationException("You are not authorized");
+        }
+
         return personalPlanRepository.findByClientIdAndIsActiveTrue(clientId)
-                .map(personalPlanDTOMapper::mapEntityToDTO);
+                .map(personalPlanDTOMapper::mapEntityToDTO).orElseThrow(
+                        () -> new NoActivePlanException("Client do not have active plan")
+                );
     }
 
-    public void deactivatePlan(Long planId) {
+    public void deactivatePlan(Long planId, User user) {
         PersonalPlan plan = personalPlanRepository.findById(planId).orElseThrow(
                 () -> new EntityNotFoundException("Plan not found")
         );
+
+        Client client = plan.getClient();
+        if (!Objects.equals(client.getCoach().getEmail(), user.getEmail()) && !user.getRole().equals(Role.ADMIN)) {
+            throw new AuthorizationException("You are not authorized");
+        }
+
         plan.setActive(false);
         plan.setEndDate(LocalDate.now());
+
+        for (Mesocycle mesocycle: plan.getMesocycles()) {
+            for (Microcycle microcycle: mesocycle.getMicrocycles()) {
+                for (WorkoutUnit workoutUnit: microcycle.getWorkoutUnits()) {
+                    if (!workoutUnit.isIndividual()) {
+                        coachSchedulerService.removeWorkout(workoutUnit.getId());
+                    }
+                }
+            }
+        }
+
         personalPlanRepository.save(plan);
     }
 
-    public PersonalPlanDTO createPersonalPlan(PersonalPlanDTO planDTO) {
+    public PersonalPlanDTO createPersonalPlan(PersonalPlanDTO planDTO, User user) {
         setWorkoutDatesForPlan(planDTO);
         PersonalPlan plan = personalPlanDTOMapper.mapDTOToEntity(planDTO);
+
+        Client client = plan.getClient();
+        if (!Objects.equals(client.getCoach().getEmail(), user.getEmail()) && !user.getRole().equals(Role.ADMIN)) {
+            throw new AuthorizationException("You are not authorized");
+        }
 
         plan.getMesocycles().forEach(mesocycle -> {
             mesocycle.setId(null);
             mesocycle.getMicrocycles().forEach(microcycle -> {
                 microcycle.setId(null);
-                microcycle.getWorkouts().forEach(workout -> {
-                    workout.setId(null);
-                    workout.getWorkoutExercises().forEach(exercise -> exercise.setId(null));
-                });
+                List<Workout> workouts = new ArrayList<>();
+                microcycle.setWorkouts(workouts);
+                microcycle.getWorkoutUnits().forEach(workout ->
+                        workout.getWorkoutExercises().forEach(exercise -> exercise.setId(null)));
             });
         });
 
@@ -87,18 +136,29 @@ public class PersonalPlanService {
         return personalPlanDTOMapper.mapEntityToDTO(savedPlan);
     }
 
-    public void deletePlan(Long id) {
-        personalPlanRepository.findById(id).orElseThrow(
+    public void deletePlan(Long id, User user) {
+        PersonalPlan plan = personalPlanRepository.findById(id).orElseThrow(
                 () -> new EntityNotFoundException("Personal plan not found")
         );
+
+        Client client = plan.getClient();
+        if (!Objects.equals(client.getCoach().getEmail(), user.getEmail()) && !user.getRole().equals(Role.ADMIN)) {
+            throw new AuthorizationException("You are not authorized");
+        }
 
         personalPlanRepository.deleteById(id);
     }
 
-    public PersonalPlanDTO getPersonalPlanByWorkoutId(Long workoutId) {
-        PersonalPlan personalPlan = personalPlanRepository.findByWorkoutId(workoutId).orElseThrow(
+    public PersonalPlanDTO getPersonalPlanByWorkoutId(Long workoutId, User user) {
+        PersonalPlan personalPlan = personalPlanRepository.findByWorkoutUnitId(workoutId).orElseThrow(
                 () -> new EntityNotFoundException("Personal plan not found for the given workout ID")
         );
+
+        Client client = personalPlan.getClient();
+        if (!Objects.equals(client.getCoach().getEmail(), user.getEmail()) && !user.getRole().equals(Role.ADMIN)) {
+            throw new AuthorizationException("You are not authorized");
+        }
+
         return personalPlanDTOMapper.mapEntityToDTO(personalPlan);
     }
 
@@ -108,11 +168,20 @@ public class PersonalPlanService {
         );
 
         LocalDate startDate = personalPlan.getStartDate();
-
         Coach coach = personalPlan.getClient().getCoach();
 
         if (!Objects.equals(coach.getEmail(), user.getEmail()) && !user.getRole().equals(Role.ADMIN)) {
             throw new AuthorizationException("You are not authorized");
+        }
+
+        for (Mesocycle mesocycle : personalPlan.getMesocycles()) {
+            for (Microcycle microcycle : mesocycle.getMicrocycles()) {
+                for (WorkoutUnit workoutUnit : microcycle.getWorkoutUnits()) {
+                    if (!workoutUnit.isIndividual()) {
+                        coachSchedulerService.removeWorkout(workoutUnit.getId());
+                    }
+                }
+            }
         }
 
         personalPlanRepository.delete(personalPlan);
@@ -134,31 +203,17 @@ public class PersonalPlanService {
         for (MesocycleDTO mesocycle : personalPlanDTO.getMesocycles()) {
             for (MicrocycleDTO microcycle : mesocycle.getMicrocycles()) {
                 Integer microcycleLength = microcycle.getLength();
+                for (WorkoutUnitDTO workout : microcycle.getWorkoutUnits()) {
 
-                for (WorkoutDTO workout : microcycle.getWorkouts()) {
-                    List<WorkoutDateDTO> workoutDates = createWorkoutDates(workout, currentDate, dayCount);
+                    LocalDate date = currentDate.plusDays(dayCount + (workout.getMicrocycleDay() - 1));
+                    LocalDateTime dateTime = LocalDateTime.of(date, LocalTime.of(0, 0));
 
-                    workout.setDates(workoutDates);
+                    workout.setDate(dateTime);
+                    workout.setDuration(60);
+                    workout.setIndividual(true);
                 }
                 dayCount += microcycleLength;
             }
         }
-    }
-
-    private static List<WorkoutDateDTO> createWorkoutDates(WorkoutDTO workout, LocalDate currentDate, int dayCount) {
-        List<WorkoutDateDTO> workoutDates = new ArrayList<>();
-
-        for (Integer day : workout.getDays()) {
-            WorkoutDateDTO workoutDate = new WorkoutDateDTO();
-            LocalDate date = currentDate.plusDays(dayCount + (day - 1));
-            LocalDateTime dateTime = LocalDateTime.of(date, LocalTime.of(0, 0));
-
-            workoutDate.setDate(dateTime);
-            workoutDate.setDuration(60);
-            workoutDate.setIndividual(true);
-
-            workoutDates.add(workoutDate);
-        }
-        return workoutDates;
     }
 }
